Naprawa szybkości gry
Dodanie obsługi FPSów do pętli w metodzie run klasy Scene.

+++ Engine/DynamicEntity.h
@@ -14,7 +14,7 @@ protected:
     int isMovingLeft;
     int isMovingUp;
     int isMovingDown;
-    int velocity;
+    double velocity;
     int acceleration;
 
 public:
